+++ CHANGELOG.md
@@ -1,3 +1,92 @@
+## Changes in 0.24.0
+
+### Fixes
+
+* Use the association primary key when importing. Thanks to @dpogue via \#512.
+* Allow association ids to be updated. Thanks to @Aristat via \#515.
+
+## Changes in 0.23.0
+
+### New Features
+
+* Rename `import` method to `bulk_import` and alias to `import`. Thanks
+  to @itay-grudev, @jkowens via \#498.
+* Increment lock_version on duplicate key update. Thanks to @aimerald
+  via \#500.
+
+### Fixes
+
+* Fix import_without_validations_or_callbacks exception if array is empty.
+  Thanks to @doloopwhile via \#508.
+
+## Changes in 0.22.0
+
+### New Features
+
+* Add support for importing hashes thru a has many association. Thanks
+  to @jkowens via \#483.
+
+### Fixes
+
+* Fix validation logic for recursive import. For those on Rails 5.0 and 5.1,
+  this change requires models with polymorphic associations to specify the `inverse_of`
+  argument (See issue #495). Thanks to @eric-simonton-sama, @jkowens via
+  \#489.
+
+## Changes in 0.21.0
+
+### New Features
+
+* Allow SQL subqueries (objects that respond to .to_sql) to be passed as values. Thanks
+  to @jalada, @jkowens via \#471
+* Raise an ArgumentError when importing an array of hashes if any of the
+  hash objects have different keys. Thanks to @mbell697 via \#465.
+
+### Fixes
+
+* Fix issue loading incorrect foreign key value when syncing belongs_to
+  associations with custom foreign key columns. Thanks to @marcgreenstock, @jkowens via \#470.
+* Fix issue importing models with polymorphic belongs_to associations.
+  Thanks to @zorab47, @jkowens via \#476.
+* Fix issue importing STI models with ActiveRecord 4.0. Thanks to
+  @kazuki-st, @jkowens via \#478.
+
+## Changes in 0.20.2
+
+### Fixes
+
+* Unscope model when synchronizing with database. Thanks to @indigoviolet via \#455.
+
+## Changes in 0.20.1
+
+### Fixes
+
+* Prevent :on_duplicate_key_update args from being modified. Thanks to @joshuamcginnis, @jkowens via \#451.
+
+## Changes in 0.20.0
+
+### New Features
+
+* Allow returning columns to be specified for PostgreSQL. Thanks to
+  @tjwp via \#433.
+
+### Fixes
+
+* Fixes an issue when bypassing uniqueness validators. Thanks to @vmaxv via \#444.
+* For AR < 4.2, prevent type casting for binary columns on Postgresql. Thanks to @mwalsher via \#446.
+* Fix issue logging class name on import. Thanks to @sophylee, @jkowens via \#447.
+* Copy belongs_to association id to foreign key column before importing. Thanks to @jkowens via \#448.
+* Reset model instance on validate. Thanks to @vmaxv via \#449.
+
+## Changes in 0.19.1
+
+### Fixes
+
+* Fix a regression where models weren't properly being marked clean. Thanks to @tjwp via \#434.
+* Raise ActiveRecord::Import::ValueSetTooLargeError when a record being inserted exceeds the
+  `max_allowed_packet` for MySQL. Thanks to @saizai, @jkowens via \#437.
+* Fix issue concatenating column names array with primary key. Thanks to @keeguon via \#440.
+
 ## Changes in 0.19.0
 
 ### New Features